separate usecase api service files
